Renamed all components.tsx to index.tsx (╯‵□′)╯︵┻━┻
