+++ backhaul-xcien/src/pages/IndexPage/index.tsx
@@ -2,8 +2,7 @@ import { FC, useState } from "react";
 import Graph from "@/components/graph";
 import { useCyContext } from "@/hooks/useCyContext";
 import { getRandomPosition } from "./utils";
-import Sidebar from "@/components/Sidebar/Sidebar";
-import SidebarToggleButton from "@/components/Sidebar/SidebarToggleButton/SidebarToggleButton";
+import Sidebar from "@/components/Sidebar";
 import NodeModal from "@/components/modal/modal";
 import LinkModal from "@/components/modal2/modal2";
 import FloatingActionBar from "@/components/FloatingActionBar/FloatingActionBar";
